pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=70.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qoder-test-platform"
version = "0.1.0"
description = "AI-Native Test Platform - CLI and Agent framework"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "QualityAI Team"}
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "python-frontmatter>=1.0.0",
    "markdown>=3.5.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.0.20",
    "jinja2>=3.0.0",
    "pyyaml>=6.0",
    "tomli>=2.0.0;python_version<'3.11'",
]

[project.optional-dependencies]
ui = [
    "playwright>=1.45.0",
    "pytest>=8.0.0",
    "pytest-json-report>=1.5.0",
]
api = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-json-report>=1.5.0",
    "jsonschema>=4.22.0",
]
kb = [
    "qdrant-client>=1.8.0",
    "fastembed>=0.2.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
aq = "qoder.cli.app:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["qoder*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
